add update functionality

+++ src/index.js
@@ -10,6 +10,12 @@ document.addEventListener('DOMContentLoaded', () => {
     createFormHandler(e)
     e.target.reset()
     })
+    // const visionContainer = document.querySelector('#vision-container')
+    // visionContainer.addEventListener('click', e => {
+    //     const id = parseInt(e.target.dataset.id)
+    //     const vision = Vision.findById(id)
+    //     document.querySelector('#update-vision').addEventListener('submit', e => updateFormHandler(e))
+    // })
 })
 
 function getVisions() {
@@ -61,14 +67,86 @@ function postFetch(title, description, image_url, theme_id) {
 function mountEditDestroy() {
     const visionContainer = document.querySelector('#vision-container')
     visionContainer.addEventListener('click', e => {
-        deleteVision(e)
-        e.target.parentElement.remove()
+        if (e.target.className === "edit-button") {
+            const currentImage = e.target.parentElement.querySelector('img')
+            const currentTitle = e.target.parentElement.querySelector('h3')
+            const currentDescription = e.target.parentElement.querySelector('h4')
+            const currentTheme = e.target.parentElement.querySelector('p')
+            const id = e.target.dataset.id
 
-        // document.querySelector('#update-vision').innerHTML = vision.renderUpdateForm()
+            const image = document.querySelector('#input-url')
+            const title = document.querySelector('#input-title')
+            const description = document.querySelector('#input-description')
+            const theme = document.querySelector('#themes')
+            const submit = document.querySelector('#create-button')
+
+            image.value = currentImage.innerText
+            title.value = currentTitle.innerText
+            description.value = currentDescription.innerText
+            theme.value = currentTheme.innerText
+
+            submit.value = "Edit Vision"
+
+            const form = document.querySelector('#create-vision-form')
+            form.dataset.action = "update"
+
+
+
+
+            // fetch(`${visionsURL}/${e.target.dataset.id}`), {
+
+            //         method: "PATCH",
+            //         headers: {
+            //             'Content-Type': 'application/json',
+            //             'Accept': 'application/json'
+            //         },
+            //         body: JSON.stringify({vision: visionObj})
+               
+            // })
+            
+                //     .then(resp => resp.json())
+                //     .then(vision => {
+                //         console.log(vision)
+            
+        }
         
+        
+        else if (e.target.className === "delete-button") {
+            deleteVision(e)
+            e.target.parentElement.remove()
+        // document.querySelector('#update-vision').innerHTML = vision.renderUpdateForm()
+        }
     })
-
+    
 }
+
+// function updateFormHandler(e) {
+//     e.preventDefault()
+//     const id = parseInt(e.target.dataset.id);
+//     const vision = Vision.findById(id);
+//     const title = e.target.querySelector('#input-title').value;
+//     const description = e.target.querySelector('#input-description').value;
+//     const image_url = e.target.querySelector('#input-url').value;
+//     const theme_id = parseInt(e.target.querySelector('#themes').value);
+//     patchVision(vision, title, description, image_url, theme_id)
+// }
+
+// function patchVision(vision, title, description, image_url, theme_id) {
+//     const bodyJSON = { title, description, image_url, theme_id }
+//     fetch(`${visionsURL}/${vision.id})`, {
+//         method: "PATCH",
+//         headers: {
+//             'Content-Type': 'application/json',
+//             'Accept': 'application/json'
+//         },
+//         body: JSON.stringify(bodyJSON),
+//         })
+//         .then(resp => resp.json())
+//         .then(updatedVision => {
+//             console.log(updatedVision)
+//         })
+// }
+
 
 function deleteVision(e) {
     fetch(`${visionsURL}/${e.target.dataset.id}`, {
